Extract activation to a new module.
